leave message

+++ src/components/chat/ChatHead.tsx
@@ -109,27 +109,36 @@ const ChatHead = ({ isNew }: { isNew?: boolean }) => {
     try {
       switch (configId) {
         case 2:
-          const res = await change(
-            `chat/member/${
-              currentChatProfileDetails?.chatMembers?.find(
-                (item) => item?.user?.id === user?.id
-              )?.id
-            }`,
-            {
-              method: "DELETE",
-              BASE_URL,
-            }
-          );
-
-          if (res?.status !== 201) {
-            Swal.fire(
-              "Error",
-              res?.results?.msg || "Something went wrong!",
-              "error"
+          if (
+            !currentChatProfileDetails?.chatMembers?.find(
+              (item) => item?.user?.id === user?.id
+            )?.isPastMember
+          ) {
+            const res = await change(
+              `chat/member/${
+                currentChatProfileDetails?.chatMembers?.find(
+                  (item) => item?.user?.id === user?.id
+                )?.id
+              }`,
+              {
+                method: "DELETE",
+                BASE_URL,
+              }
             );
+
+            console.log("remove");
+            if (res?.status !== 201) {
+              Swal.fire(
+                "Error",
+                res?.results?.msg || "Something went wrong!",
+                "error"
+              );
+            }
+            revalidateCurrentChat(selectedChatId);
+            selectedChatId && revalidateChatProfileDetails(selectedChatId);
+            break;
           }
-          revalidateCurrentChat(selectedChatId);
-          selectedChatId && revalidateChatProfileDetails(selectedChatId);
+          Swal.fire("Info", "You leave this group already", "info");
           break;
         case 1:
           const blockUser = await change(`chat/blocked/${selectedChatId}`, {
